refactor: Cleanup custom view
- All custom view now extends CustomView class
- Cleaned up view methods to ease reading

+++ app/src/main/java/com/paradoxcat/waveviewer/util/TimeConverter.kt
@@ -18,7 +18,7 @@ object TimeConverter {
      * @param currentMilliseconds current timestamp of the audio file
      * @param totalMilliseconds total duration of the audio file
      */
-    fun millisecondsToProgress(currentMilliseconds: Long, totalMilliseconds: Int): Int {
+    fun millisecondsToProgress(currentMilliseconds: Long, totalMilliseconds: Long): Int {
         return (currentMilliseconds * MAX_PROGRESS_VALUE / totalMilliseconds).toInt()
     }
